JAMES-2393 Introduce dedicated modules for event sourcing
This intend to ease future usage, external usage as well as allow a clear view.
