+++ backends-common/eventsourcing-memory/src/test/java/org/apache/james/eventsourcing/InMemoryEventSourcingSystemTest.java
