+++ backends-common/eventsourcing/src/main/java/org/apache/james/eventsourcing/Subscriber.java
