+++ backends-common/eventsourcing/src/test/java/org/apache/james/eventsourcing/DataCollectorSubscriber.java
