+++ mailbox/plugin/quota-mailing-cassandra/src/main/java/org/apache/james/mailbox/quota/cassandra/dto/QuotaThresholdChangedEventDTO.java
@@ -20,8 +20,8 @@
 package org.apache.james.mailbox.quota.cassandra.dto;
 
 import org.apache.james.core.User;
+import org.apache.james.evensourcing.cassandra.dto.EventDTO;
 import org.apache.james.eventsourcing.EventId;
-import org.apache.james.eventsourcing.cassandra.dto.EventDTO;
 import org.apache.james.mailbox.quota.mailing.aggregates.UserQuotaThresholds;
 import org.apache.james.mailbox.quota.mailing.events.QuotaThresholdChangedEvent;
