+++ mailbox/plugin/quota-mailing-cassandra/src/test/java/org/apache/james/mailbox/quota/cassandra/listeners/CassandraQuotaMailingListenersIntegrationTest.java
@@ -19,7 +19,7 @@
 
 package org.apache.james.mailbox.quota.cassandra.listeners;
 
-import org.apache.james.eventsourcing.cassandra.CassandraEventStoreExtension;
+import org.apache.james.mailbox.quota.cassandra.CassandraEventStoreExtension;
 import org.apache.james.mailbox.quota.mailing.listeners.QuotaThresholdMailingIntegrationTest;
 import org.junit.jupiter.api.extension.ExtendWith;
